label-relation/requirements.txt
Flask==2.3.3
neo4j==5.15.0
spacy==3.7.2
flask-cors==4.0.0
langchain==0.1.0
torch==2.1.0
torch-geometric==2.4.0
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2
APScheduler==3.10.4
websockets==11.0.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0 